# src/main/resources/openapi/healthcare-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Healthcare OPD Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/patients/profile": {
      "get": {
        "tags": ["Patients"],
        "operationId": "getPatientProfile",
        "summary": "Get the current patient's own profile",
        "description": "PATIENT only.",
        "responses": {
          "200": {
            "description": "Patient profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Patients"],
        "operationId": "updatePatientProfile",
        "summary": "Update the current patient's own profile",
        "description": "PATIENT only.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatientUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient profile updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientResponse" }
              }
            }
          }
        }
      }
    },
    "/api/patients/{id}": {
      "get": {
        "tags": ["Patients"],
        "operationId": "getPatientById",
        "summary": "Get a patient by id",
        "description": "ADMIN or DOCTOR only.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Patient detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientResponse" }
              }
            }
          }
        }
      }
    },
    "/api/patients/search": {
      "get": {
        "tags": ["Patients"],
        "operationId": "searchPatients",
        "summary": "Search patients",
        "description": "ADMIN only.",
        "parameters": [
          { "name": "query", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated patient search results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PatientSummaryPagedResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/pmr/patient/{patientId}/entries": {
      "get": {
        "tags": ["PMR"],
        "operationId": "getPmrEntries",
        "summary": "List PMR entries for a patient",
        "description": "PATIENT (own only). DOCTOR access requires an active appointment window with the patient. ADMIN has no PMR access.",
        "parameters": [
          { "name": "patientId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "hospitalId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated PMR entries",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PmrEntryPagedResponse" }
              }
            }
          }
        }
      }
    },
    "/api/pmr/patient/{patientId}/entry": {
      "post": {
        "tags": ["PMR"],
        "operationId": "addPmrEntry",
        "summary": "Add a PMR entry for a patient",
        "description": "DOCTOR only, and only during an active appointment window with the patient.",
        "parameters": [
          { "name": "patientId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "hospitalId", "in": "query", "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PmrEntryRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PMR entry created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PmrEntryResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/documents/upload": {
      "post": {
        "tags": ["Documents"],
        "operationId": "uploadDocument",
        "summary": "Upload a document",
        "description": "PATIENT may upload PREVIOUS_RECORD type only (for themselves). ADMIN may upload any type for their hospital.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file", "patientId", "hospitalId", "documentType"],
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "patientId": { "type": "integer", "format": "int64" },
                  "hospitalId": { "type": "integer", "format": "int64" },
                  "appointmentId": { "type": "integer", "format": "int64" },
                  "documentType": { "type": "string", "enum": ["BILL","PRESCRIPTION","REPORT","DOCTOR_NOTE","PREVIOUS_RECORD"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document uploaded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentResponse" }
              }
            }
          }
        }
      }
    },
    "/api/documents/patient/{patientId}": {
      "get": {
        "tags": ["Documents"],
        "operationId": "getPatientDocuments",
        "summary": "List documents for a patient",
        "description": "PATIENT sees own visible documents. ADMIN sees all for their hospital. DOCTOR sees patient-uploaded documents within the appointment time window.",
        "parameters": [
          { "name": "patientId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "documentType", "in": "query", "schema": { "type": "string" } },
          { "name": "appointmentId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Document list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentSummaryPagedResponse" }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}/download": {
      "get": {
        "tags": ["Documents"],
        "operationId": "getDocumentDownloadUrl",
        "summary": "Get a SAS download URL for a document",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "SAS download URL",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DownloadUrlResponse" }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}": {
      "delete": {
        "tags": ["Documents"],
        "operationId": "deleteDocument",
        "summary": "Soft-delete a document (ADMIN only)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "Document removed" }
        }
      }
    },
    "/api/notifications/my": {
      "get": {
        "tags": ["Notifications"],
        "operationId": "getMyNotifications",
        "summary": "Get notifications for the current user",
        "parameters": [
          { "name": "isRead", "in": "query", "schema": { "type": "boolean" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated notifications",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NotificationPagedResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/{id}/read": {
      "patch": {
        "tags": ["Notifications"],
        "operationId": "markNotificationRead",
        "summary": "Mark a single notification as read",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Notification marked read",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MarkReadResponse" }
              }
            }
          }
        }
      }
    },
    "/api/notifications": {
      "post": {
        "tags": ["Notifications"],
        "operationId": "sendNotification",
        "summary": "Send a notification (ADMIN only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NotificationRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NotificationSentResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PmrEntryRequest": {
        "type": "object",
        "description": "Request payload to add a PMR entry.",
        "properties": {
          "appointmentId": {
            "type": "integer",
            "format": "int64",
            "description": "Linked appointment identifier.",
            "example": 501
          },
          "entryDate": {
            "type": "string",
            "format": "date",
            "description": "Date of the PMR entry.",
            "example": "2026-07-01"
          },
          "diagnosis": {
            "type": "string",
            "description": "Diagnosis notes.",
            "example": "Migraine without aura"
          },
          "symptoms": {
            "type": "string",
            "description": "Observed symptoms.",
            "example": "Headache, nausea"
          },
          "treatmentPlan": {
            "type": "string",
            "description": "Treatment plan recommended by the doctor.",
            "example": "Prescribe triptans and hydration"
          },
          "doctorNotes": {
            "type": "string",
            "description": "Additional doctor notes.",
            "example": "Review after two weeks"
          }
        },
        "required": [
          "appointmentId",
          "entryDate"
        ]
      },
      "PmrEntryResponse": {
        "type": "object",
        "description": "PMR entry details.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "PMR entry identifier.",
            "example": 801
          },
          "doctorId": {
            "type": "integer",
            "format": "int64",
            "description": "Doctor identifier.",
            "example": 301
          },
          "doctorName": {
            "type": "string",
            "description": "Doctor name.",
            "example": "Dr. Meera Shah"
          },
          "appointmentId": {
            "type": "integer",
            "format": "int64",
            "description": "Linked appointment identifier.",
            "example": 501
          },
          "entryDate": {
            "type": "string",
            "format": "date",
            "description": "PMR entry date.",
            "example": "2026-07-01"
          },
          "diagnosis": {
            "type": "string",
            "description": "Diagnosis details.",
            "example": "Migraine without aura"
          },
          "symptoms": {
            "type": "string",
            "description": "Symptoms recorded in the PMR.",
            "example": "Headache, nausea"
          },
          "treatmentPlan": {
            "type": "string",
            "description": "Treatment plan.",
            "example": "Prescribe triptans and hydration"
          },
          "doctorNotes": {
            "type": "string",
            "description": "Doctor notes.",
            "example": "Review after two weeks"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the PMR entry was created.",
            "example": "2026-07-01T09:45:00Z"
          }
        }
      },
      "PmrEntryPagedResponse": {
        "type": "object",
        "description": "Paginated list of PMR entries.",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PmrEntryResponse" }
          },
          "page": { "type": "integer", "example": 0 },
          "size": { "type": "integer", "example": 20 },
          "totalElements": { "type": "integer", "format": "int64", "example": 42 },
          "totalPages": { "type": "integer", "example": 3 }
        }
      },
...
      "PatientResponse": {
        "type": "object",
        "description": "Patient detail response.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Patient identifier.",
            "example": 201
          },
          "name": {
            "type": "string",
            "description": "Patient name.",
            "example": "Asha Patil"
          },
          "age": {
            "type": "integer",
            "description": "Patient age in years, computed from dob.",
            "example": 34,
            "readOnly": true
          },
          "gender": {
            "type": "string",
            "description": "Patient gender.",
            "example": "FEMALE"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "description": "Patient date of birth.",
            "example": "1992-08-15"
          },
          "mobileNumber": {
            "type": "string",
            "description": "Patient mobile number.",
            "example": "9876543210"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Patient email address.",
            "example": "asha.patil@example.com"
          },
          "address": {
            "type": "string",
            "description": "Patient residential address.",
            "example": "12 MG Road, Pune, MH"
          },
          "emergencyContactNumber": {
            "type": "string",
            "description": "Emergency contact mobile number.",
            "example": "9876500000"
          },
          "bloodGroup": {
            "type": "string",
            "enum": ["A+", "A-", "B+", "B-", "AB+", "AB-", "O+", "O-"],
            "description": "Patient blood group.",
            "example": "O+"
          },
          "permanentIllness": {
            "type": "string",
            "description": "Long-term / permanent illness details, if any.",
            "example": "Type 2 Diabetes"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the patient was created.",
            "example": "2026-06-27T11:30:00Z"
          }
        }
      },
      "PatientUpdateRequest": {
        "type": "object",
        "description": "Request payload to update patient profile details.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated patient name.",
            "example": "Asha P. Patil"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "description": "Updated patient date of birth.",
            "example": "1991-08-15"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Updated patient email address.",
            "example": "asha.new@example.com"
          },
          "address": {
            "type": "string",
            "description": "Updated patient residential address.",
            "example": "12 MG Road, Pune, MH"
          },
          "emergencyContactNumber": {
            "type": "string",
            "pattern": "^[0-9]{10}$",
            "description": "Updated emergency contact 10-digit mobile number.",
            "example": "9876500000"
          },
          "bloodGroup": {
            "type": "string",
            "enum": ["A+", "A-", "B+", "B-", "AB+", "AB-", "O+", "O-"],
            "description": "Updated patient blood group.",
            "example": "O+"
          },
          "permanentIllness": {
            "type": "string",
            "description": "Updated long-term / permanent illness details.",
            "example": "Type 2 Diabetes"
          }
        }
      },
...
      "PatientSummaryPagedResponse": {
        "type": "object",
        "description": "Paginated list of patient summaries.",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PatientSummaryResponse" }
          },
          "page": { "type": "integer", "example": 0 },
          "size": { "type": "integer", "example": 20 },
          "totalElements": { "type": "integer", "format": "int64", "example": 42 },
          "totalPages": { "type": "integer", "example": 3 }
        }
      },
      "DocumentResponse": {
        "type": "object",
        "description": "Detailed uploaded document response.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Document identifier.",
            "example": 1001
          },
          "patientId": {
            "type": "integer",
            "format": "int64",
            "description": "Patient identifier.",
            "example": 201
          },
          "hospitalId": {
            "type": "integer",
            "format": "int64",
            "description": "Hospital identifier.",
            "example": 12
          },
          "appointmentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Related appointment identifier when applicable.",
            "example": 501
          },
          "documentType": {
            "type": "string",
            "description": "Document classification.",
            "example": "PREVIOUS_RECORD"
          },
          "originalFileName": {
            "type": "string",
            "description": "Original uploaded file name.",
            "example": "blood-report.pdf"
          },
          "uploadedByRole": {
            "type": "string",
            "description": "Role that uploaded the document.",
            "example": "PATIENT"
          },
          "isVisibleToPatient": {
            "type": "boolean",
            "description": "Whether the patient can view the document.",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the document was created.",
            "example": "2026-06-27T11:30:00Z"
          }
        }
      },
...
      "DocumentSummaryPagedResponse": {
        "type": "object",
        "description": "Paginated list of document summaries.",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DocumentSummaryResponse" }
          },
          "page": { "type": "integer", "example": 0 },
          "size": { "type": "integer", "example": 20 },
          "totalElements": { "type": "integer", "format": "int64", "example": 42 },
          "totalPages": { "type": "integer", "example": 3 }
        }
      },
      "DownloadUrlResponse": {
        "type": "object",
        "description": "Time-limited download URL for a document.",
        "properties": {
          "downloadUrl": {
            "type": "string",
            "format": "uri",
            "description": "SAS URL for downloading the document.",
            "example": "https://storage.example.com/container/blob.pdf?sv=..."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Offset timestamp when the URL expires.",
            "example": "2026-06-27T12:30:00+05:30"
          }
        }
      },
...
      "NotificationPagedResponse": {
        "type": "object",
        "description": "Paginated list of notifications.",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/NotificationResponse" }
          },
          "page": { "type": "integer", "example": 0 },
          "size": { "type": "integer", "example": 20 },
          "totalElements": { "type": "integer", "format": "int64", "example": 42 },
          "totalPages": { "type": "integer", "example": 3 }
        }
      },
...
      "MarkReadResponse": {
        "type": "object",
        "description": "Response confirming a notification read-state update.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Notification identifier.",
            "example": 1101
          },
          "isRead": {
            "type": "boolean",
            "description": "Updated read status.",
            "example": true
          }
        }
      },
...
    }
  }
}